# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo API",
    "version": "0.1.0",
    "description": "HTTP API for the Igloo self-hosted media center. Most JSON endpoints return a shared envelope with error, optional message, and optional data fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Igloo server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/movies/{id}/hls/{profile}/playlist.m3u8": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Get an HLS playlist for a movie",
        "operationId": "hlsManifest",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          },
          {
            "$ref": "#/components/parameters/HLSProfilePath"
          },
          {
            "$ref": "#/components/parameters/AudioTrackQuery"
          },
          {
            "$ref": "#/components/parameters/StartQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HLSPlaylistResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/movies/{id}/hls/{profile}/{filename}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Get an HLS initialization file or media segment for a movie",
        "operationId": "hlsSegment",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          },
          {
            "$ref": "#/components/parameters/HLSProfilePath"
          },
          {
            "$ref": "#/components/parameters/HLSFilenamePath"
          },
          {
            "$ref": "#/components/parameters/AudioTrackQuery"
          },
          {
            "$ref": "#/components/parameters/StartQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HLSSegmentResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/settings/playback": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get playback preferences and available profiles",
        "operationId": "getPlaybackSettings",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PlaybackSettingsResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update playback preferences",
        "operationId": "updatePlaybackSettings",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdatePlaybackSettingsRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UpdatePlaybackSettingsResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/music/musicians/": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "List musicians alphabetically",
        "operationId": "getMusiciansAlphabetical",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/PerPageQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MusiciansResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/music/musicians/{id}": {
      "get": {
        "tags": [
          "Music"
        ],
        "summary": "Get musician details",
        "operationId": "getMusicianDetails",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MusicianDetailsResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Server-side session cookie set by POST /api/auth/login."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
      "HLSProfilePath": {
        "name": "profile",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/HLSProfile"
        }
      },
      "HLSFilenamePath": {
        "name": "filename",
        "in": "path",
        "required": true,
        "description": "HLS asset name. Valid values are init.mp4 or segment_N.m4s.",
        "schema": {
          "type": "string",
          "pattern": "^(init\\.mp4|segment_[0-9]+\\.m4s)$"
        }
      },
...
      "PageQuery": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PerPageQuery": {
        "name": "per_page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
      "AudioTrackQuery": {
        "name": "audio_track",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "StartQuery": {
        "name": "start",
        "in": "query",
        "required": false,
        "description": "Session start time in seconds.",
        "schema": {
          "type": "number",
          "minimum": 0,
          "default": 0
        }
      },
...
    },
    "requestBodies": {
...
      "UpdatePlaybackSettingsRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdatePlaybackSettingsRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "PlaybackSettingsResponse": {
        "description": "Playback settings response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PlaybackSettingsEnvelope"
            }
          }
        }
      },
      "UpdatePlaybackSettingsResponse": {
        "description": "Updated playback settings response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdatePlaybackSettingsEnvelope"
            }
          }
        }
      },
...
      "HLSPlaylistResponse": {
        "description": "HLS media playlist.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          }
        },
        "content": {
          "application/vnd.apple.mpegurl": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "HLSSegmentResponse": {
        "description": "Fragmented MP4 initialization file or media segment.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          }
        },
        "content": {
          "video/mp4": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
      "MusiciansResponse": {
        "description": "Paginated musicians response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MusiciansEnvelope"
            }
          }
        }
      },
      "MusicianDetailsResponse": {
        "description": "Musician details response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MusicianDetailsEnvelope"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Generated media segment is not ready yet.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "JsonSuccess": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "HLSProfile": {
        "type": "string",
        "enum": [
          "remux",
          "2160p_16mbps",
          "1080p_8mbps",
          "1080p_6mbps",
          "1080p_4mbps",
          "720p_3mbps"
        ]
      },
...
      "UpdatePlaybackSettingsRequest": {
        "type": "object",
        "properties": {
          "preferred_profile": {
            "type": [
              "string",
              "null"
            ],
            "description": "Allowed transcode profile except remux."
          },
          "download_mbps": {
            "type": [
              "number",
              "null"
            ],
            "exclusiveMinimum": 0,
            "exclusiveMaximum": 10000
          },
          "preferred_audio_language": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[a-z]{2,3}$"
          },
          "preferred_subtitle_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "Two- or three-letter language code, or off."
          }
        },
        "additionalProperties": false
      },
...
      "PlaybackSettingsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "settings"
                ],
                "properties": {
                  "settings": {
                    "$ref": "#/components/schemas/PlaybackSettings"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "UpdatePlaybackSettingsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "settings"
                ],
                "properties": {
                  "settings": {
                    "$ref": "#/components/schemas/UpdatePlaybackSettings"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
      "MusiciansEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "musicians",
                  "total",
                  "page",
                  "per_page",
                  "total_pages"
                ],
                "properties": {
                  "musicians": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SimpleMusician"
                    }
                  },
                  "total": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "page": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "per_page": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "total_pages": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "MusicianDetailsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "musician",
                  "albums",
                  "tracks",
                  "genres",
                  "total_duration"
                ],
                "properties": {
                  "musician": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "albums": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "tracks": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "genres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "total_duration": {
                    "type": "number"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
    }
  }
}